openanalytics-main/requirements.txt
# Core
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
httpx>=0.25.0

# HTML parsing
beautifulsoup4>=4.12.0
lxml>=4.9.0

# AI
google-genai>=0.2.0
python-dotenv>=1.0.0

# Playwright for health check
playwright>=1.40.0
